HW_8

+++ homework_2/src/main/java/com/example/homework_2/domain/entity/MessageItem.kt
@@ -3,8 +3,8 @@ package com.example.homework_2.domain.entity
 import com.example.homework_2.data.network.model.chat.reaction.ReactionResponse
 
 data class MessageItem(
-    val id: Long,
-    val userId: Long? = null,
+    val id: Int,
+    val userId: Int? = null,
     val userFullName: String? = null,
     val topicName: String? = null,
     val avatarUrl: String? = null,
